requirements.txt
copier
pytest
pytest-venv
pytest-xdist
black
mkdocs-material
mkdocs-git-revision-date-localized-plugin
mkdocs-literate-nav
mkdocs-macros-plugin
git+https://github.com/jannismain/mkdocs-macros-includex
mkdocs-exclude
mkdocs-autorefs
